+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to `laravel-rajabiller` will be documented in this file
 
+## [2.0.0] - 2021-10-28
+- Drop support for PHP7.4
+- Add Credit Card, Transfer & Buy method
+- Add process() to RbItem Model to consolidate processing
+- Change all methods to return json object instead of `ResponseInterface`
+- Update dependencies
+
 ## [1.0.3] - 2021-10-27
 - New function getItem() in Rajabiller.php